Add support for new-style dupdate zip archives
When given a zip file, dupdate unpacks only the run file, which is then
called with the archive as argument. To use other files in the archive,
the run file can then unpack such files to stdout and pipe them directly
to a command, thus not requiring intermediate filesystem space for them.
Or it could just star by unpacking the entire archive and do as with the
old-style dupdate tar archive.

Old-style dupdate tar archives are still supported and are processed as
usual.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>