Split out dupdate image processing from inotify event loop
This commit splits out the dupdate image processing functionality into a
stand-alone dupdate command, and the inotify event processing into a
separate inotifyd command.

The inotifyd command is implemented as a drop-in replacement for the
busybox inotifyd applet.  If you are using busybox, you should consider
using that instead.

A dupdate-inotifyd-agent shell script provides the glue needed to
link together inotifyd and the new dupdate command to provide the same
functionality as the old dupdate daemon did.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>